Some more minor fixes

+++ src/config/const.hpp
@@ -58,7 +58,7 @@
 /**
  * EXP definition type
  */
-typedef unsigned long long t_exp;
+typedef uint64 t_exp;
 #if PACKETVER >= 20170830
 	const t_exp EXP_MAX = INT64_MAX;
 #else
